+++ spec/rubocop/cop/rails/scope_args_spec.rb
@@ -1,10 +1,14 @@
 # frozen_string_literal: true
 
 RSpec.describe RuboCop::Cop::Rails::ScopeArgs, :config do
-  it 'registers an offense a scope with a method arg' do
+  it 'registers and corrects an offense a scope with a method arg' do
     expect_offense(<<~RUBY)
       scope :active, where(active: true)
                      ^^^^^^^^^^^^^^^^^^^ Use `lambda`/`proc` instead of a plain method call.
+    RUBY
+
+    expect_correction(<<~RUBY)
+      scope :active, -> { where(active: true) }
     RUBY
   end